requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
asyncer==0.0.17
attrs==26.1.0
audioop-lts==0.2.2
bcrypt==5.0.0
bidict==0.23.1
build==1.4.2
certifi==2026.2.25
cffi==2.0.0
chainlit==2.11.0
charset-normalizer==3.4.7
chevron==0.14.0
chromadb==1.5.7
click==8.3.2
colorama==0.4.6
cryptography==46.0.7
cuid==0.4
dataclasses-json==0.6.7
Deprecated==1.3.1
distro==1.9.0
durationpy==0.10
fastapi==0.135.3
filelock==3.25.2
filetype==1.2.0
flatbuffers==25.12.19
frozenlist==1.8.0
fsspec==2026.3.0
google-auth==2.49.1
google-genai==1.71.0
googleapis-common-protos==1.74.0
grpcio==1.80.0
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.9.2
idna==3.11
importlib_metadata==8.7.1
importlib_resources==6.5.2
inflection==0.5.1
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.1.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kubernetes==35.0.0
langchain==1.2.15
langchain-chroma==1.1.0
langchain-core==1.2.28
langchain-google-genai==4.2.1
langgraph==1.1.6
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.9
langgraph-sdk==0.3.13
langsmith==0.7.29
Lazify==0.4.0
literalai==0.1.201
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
mcp==1.27.0
mdurl==0.1.2
mmh3==5.2.1
mpmath==1.3.0
multidict==6.7.1
mypy_extensions==1.1.0
nest-asyncio==1.6.0
numpy==2.4.4
oauthlib==3.3.1
onnxruntime==1.24.4
opentelemetry-api==1.40.0
opentelemetry-exporter-otlp-proto-common==1.40.0
opentelemetry-exporter-otlp-proto-grpc==1.40.0
opentelemetry-exporter-otlp-proto-http==1.40.0
opentelemetry-instrumentation==0.61b0
opentelemetry-instrumentation-agno==0.57.0
opentelemetry-instrumentation-alephalpha==0.57.0
opentelemetry-instrumentation-anthropic==0.57.0
opentelemetry-instrumentation-bedrock==0.57.0
opentelemetry-instrumentation-chromadb==0.57.0
opentelemetry-instrumentation-cohere==0.57.0
opentelemetry-instrumentation-crewai==0.57.0
opentelemetry-instrumentation-google-generativeai==0.57.0
opentelemetry-instrumentation-groq==0.57.0
opentelemetry-instrumentation-haystack==0.57.0
opentelemetry-instrumentation-lancedb==0.57.0
opentelemetry-instrumentation-langchain==0.57.0
opentelemetry-instrumentation-llamaindex==0.57.0
opentelemetry-instrumentation-logging==0.61b0
opentelemetry-instrumentation-marqo==0.57.0
opentelemetry-instrumentation-mcp==0.57.0
opentelemetry-instrumentation-milvus==0.57.0
opentelemetry-instrumentation-mistralai==0.57.0
opentelemetry-instrumentation-ollama==0.57.0
opentelemetry-instrumentation-openai==0.57.0
opentelemetry-instrumentation-openai-agents==0.57.0
opentelemetry-instrumentation-pinecone==0.57.0
opentelemetry-instrumentation-qdrant==0.57.0
opentelemetry-instrumentation-redis==0.61b0
opentelemetry-instrumentation-replicate==0.57.0
opentelemetry-instrumentation-requests==0.61b0
opentelemetry-instrumentation-sagemaker==0.57.0
opentelemetry-instrumentation-sqlalchemy==0.61b0
opentelemetry-instrumentation-threading==0.61b0
opentelemetry-instrumentation-together==0.57.0
opentelemetry-instrumentation-transformers==0.57.0
opentelemetry-instrumentation-urllib3==0.61b0
opentelemetry-instrumentation-vertexai==0.57.0
opentelemetry-instrumentation-voyageai==0.57.0
opentelemetry-instrumentation-watsonx==0.57.0
opentelemetry-instrumentation-weaviate==0.57.0
opentelemetry-instrumentation-writer==0.57.0
opentelemetry-proto==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-semantic-conventions==0.61b0
opentelemetry-semantic-conventions-ai==0.5.1
opentelemetry-util-http==0.61b0
orjson==3.11.8
ormsgpack==1.12.2
overrides==7.7.0
packaging==26.0
propcache==0.4.1
protobuf==6.33.6
pyasn1==0.6.3
pyasn1_modules==0.4.2
pybase64==1.4.3
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.20.0
PyJWT==2.12.1
PyPika==0.51.1
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-engineio==4.13.1
python-multipart==0.0.24
python-socketio==5.16.1
pywin32==311
PyYAML==6.0.3
referencing==0.37.0
requests==2.33.1
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.3.3
rpds-py==0.30.0
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
sse-starlette==3.3.4
starlette==1.0.0
sympy==1.14.0
syncer==2.0.3
tenacity==9.1.4
tokenizers==0.22.2
tomli==2.4.1
tqdm==4.67.3
traceloop-sdk==0.57.0
typer==0.24.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.44.0
watchfiles==1.1.1
websocket-client==1.9.0
websockets==16.0
wrapt==1.17.3
wsproto==1.3.2
xxhash==3.6.0
yarl==1.23.0
zipp==3.23.0
zstandard==0.25.0
